+++ tests/Feature/YamlFormTest.php
@@ -46,7 +46,7 @@ class YamlFormTest extends TestCase
             ->set("name","")
             ->set("onPullrequest", true)
             ->call('submitForm')
-            ->assertHasErrors('yaml');
+            ->assertHasErrors('name');
     }
 
     /**
@@ -62,6 +62,44 @@ class YamlFormTest extends TestCase
             ->set("onPush", false)
             ->call('submitForm')
             ->assertHasNoErrors('yaml');
+    }
+
+    /**
+     * Form Test: test on events checkboxes.
+     *
+     * @return void
+     */
+    public function test_form_submit_onevents()
+    {
+        Livewire::test(ConfiguratorForm::class)
+            ->set("name","Test")
+            ->set("manualTrigger", false)
+            ->set("onPush", false)
+            ->set("onPullrequest", false)
+            ->call('submitForm')
+            ->assertHasErrors('onEvents');
+
+        Livewire::test(ConfiguratorForm::class)
+            ->set("name","Test")
+            ->set("manualTrigger", false)
+            ->set("onPush", true)
+            ->set("onPullrequest", false)
+            ->call('submitForm')
+            ->assertHasNoErrors('onEvents');
+        Livewire::test(ConfiguratorForm::class)
+            ->set("name","Test")
+            ->set("manualTrigger", true)
+            ->set("onPush", true)
+            ->set("onPullrequest", false)
+            ->call('submitForm')
+            ->assertHasNoErrors('onEvents');
+        Livewire::test(ConfiguratorForm::class)
+            ->set("name","Test")
+            ->set("manualTrigger", true)
+            ->set("onPush", true)
+            ->set("onPullrequest", true)
+            ->call('submitForm')
+            ->assertHasNoErrors('onEvents');
     }
 
 }